Version 20240419.1
https://www.gurux.fi/release/guruxdlmscpp-202404191

+++ development/src/GXBitString.cpp
@@ -124,7 +124,7 @@ std::string CGXBitString::ToBitString(unsigned int value, unsigned int count)
             }
         }
     }
-    if (sb.width() > count)
+    if ((unsigned int) sb.width() > count)
     {
         return sb.str().substr(0, (unsigned int)(sb.width() - count));
     }
